% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_braid.R
\docType{data}
\name{StatBraid}
\alias{StatBraid}
\alias{StatBraidGlass}
\alias{StatBraidSmooth}
\alias{StatBraidContour}
\alias{braidReports-ggproto}
\title{Base ggproto classes for braidReports}
\format{
An object of class \code{StatBraid} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{StatBraidGlass} (inherits from \code{StatBraid}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatBraidSmooth} (inherits from \code{StatBraid}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.

An object of class \code{StatBraidContour} (inherits from \code{StatContour}, \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatBraid

StatBraidGlass

StatBraidSmooth

StatBraidContour
}
\description{
Base ggproto classes for braidReports
}
\seealso{
\code{\link[ggplot2:ggproto]{ggproto()}}
}
\keyword{datasets}
\keyword{internal}
