% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidReports-package.R
\docType{package}
\name{braidReports-package}
\alias{braidReports}
\alias{braidReports-package}
\title{braidReports: Visualize Combined Action Response Surfaces and Report BRAID Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to visualize combined action data in ggplot2. Also provides functions for producing full BRAID analysis reports with custom layouts and aesthetics.
}
\author{
\strong{Maintainer}: Nathaniel R. Twarog \email{nathaniel.twarog@stjude.org}

Authors:
\itemize{
  \item Anang Shelat \email{anang.shelat@stjude.org}
}

}
\keyword{internal}
