\name{es8analysis}
\docType{data}
\alias{es8analysis}
\title{Analysis of a Ewings-Sarcoma Experiment}
\description{
  This list is the results of full BRAID analysis run on the \code{\link{es8olatmz}} dataset
  from the \code{braidrm} package.  The list was output by the function
  \code{\link{runBRAIDanalysis}}, and follows the format described therein.
}
\usage{data(es8analysis)}
\format{ A list of objects produced by the function \code{\link{runBRAIDanalysis}} }
\keyword{datasets}