% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merckValues_unstable}
\alias{merckValues_unstable}
\title{Best-Fit Unstabilized Merck OPPS BRAID Values}
\format{
A data frame with 5 columns:
\describe{
\item{cell_line}{The cancer cell line which the combination was tested}
\item{drugA}{The first drug in the combination tested}
\item{drugB}{The second drug in the combination tested}
\item{kappa}{The best-fit value of the BRAID interaction parameter kappa
with no Bayesian stabilizatoin}
\item{IAE}{The index of achievable efficacy (or IAE, a BRAID measure of
combined potency) estimated for the best-fit BRAID surface}
}
}
\usage{
merckValues_unstable
}
\description{
A table of BRAID kappa and IAE values resulting from running the version
1.0.0 BRAID fitting code on the Merck oncopolypharmacology screen (OPPS),
with \emph{no} Bayesian stabilization of kappa.
}
\keyword{datasets}
